+++ src/path.util.ts
@@ -1,4 +1,3 @@
-import { JsonSchema } from './models/jsonSchema';
 import {FullDataModelType, isItemModel} from './parser/item_model';
 
 const toDataPathSegments = (path: string): Array<string> => {
@@ -64,41 +63,3 @@ export const resolveSchema = (model: FullDataModelType, path: string): FullDataM
   // }
   return resultSchema;
 };
-
-export const findAllRefs = (schema: JsonSchema, result: ReferenceSchemaMap = {}):
-  ReferenceSchemaMap => {
-  if (schema.type === 'object' && schema.properties !== undefined) {
-    Object.keys(schema.properties).forEach(key => findAllRefs(schema.properties[key], result));
-  }
-  if (schema.type === 'array' && schema.items !== undefined) {
-    // FIXME Do we want to support tupples? If so how do we render this?
-    if (Array.isArray(schema.items)) {
-      schema.items.forEach(child => findAllRefs(child, result));
-    }
-    findAllRefs(schema.items, result);
-  }
-  if (Array.isArray(schema.anyOf)) {
-      schema.anyOf.forEach(child => findAllRefs(child, result));
-  }
-  if (schema.$ref !== undefined) {
-    result[schema.$ref] = schema;
-  }
-  return result;
-};
-
-export function retrieveResolvableSchema(full: JsonSchema, reference: string): JsonSchema {
-  const child = resolveSchema(full, reference);
-  if (!isItemModel(child)) {
-    return undefined;
-  }
-  const allRefs = findAllRefs(child.schema);
-  const innerSelfReference = allRefs[reference];
-  if (innerSelfReference !== undefined) {
-    innerSelfReference.$ref = '#';
-  }
-  return child.schema;
-};
-
-export interface ReferenceSchemaMap {
-  [ref: string]: JsonSchema;
-}
