+++ test/path.util.test.ts
@@ -163,126 +163,3 @@ test('resolve $ref', t => {
       '#/properties/foos');
     t.deepEqual(result, {schema: {type: 'string'}, attributes: {}, dropPoints: {}});
 });
-test.failing('resolve $ref simple, should be moved to schema resolution', t => {
-    const schema: JsonSchema = {
-        definitions: {
-          foo: {
-            type: 'object',
-            properties: {
-              bar: {
-                type: 'array',
-                items: {
-                  $ref: '#/definitions/foo'
-                }
-              }
-            }
-          }
-        },
-        type: 'object',
-        properties: {
-          foos: {
-            type: 'array',
-            items: {
-              $ref: '#/definitions/foo'
-            }
-          }
-        }
-    } as JsonSchema;
-    const result = resolveSchema({
-      schema: schema,
-      dropPoints: {
-        foos: {
-          schema: {
-            type: 'object',
-            properties: {
-              bar: {
-                type: 'array',
-                items: {
-                  $ref: '#/definitions/foo'
-                }
-              }
-            }
-          },
-          dropPoints: {}
-        }
-      }
-    },
-      '#/properties/foos');
-    t.true(isItemModel(result));
-    t.deepEqual((<ItemModel>result).schema, {
-      type: 'object',
-      properties: {
-        bar: {
-          type: 'array',
-          items: {
-            $ref: '#'
-          }
-        }
-      }
-    });
-    t.not((<JsonSchema>schema.definitions.foo.properties.bar.items).$ref, '#');
-});
-test.failing('resolve $ref complicated, should be moved to schema resolution', t => {
-    const schema: JsonSchema = {
-        definitions: {
-          foo: {
-            type: 'object',
-            properties: {
-              bar: {
-                type: 'array',
-                items: {
-                  $ref: '#/definitions/foo2'
-                }
-              }
-            }
-          },
-          foo2: {
-            type: 'object',
-            properties: {
-              bar: {
-                type: 'array',
-                items: {
-                  $ref: '#/definitions/foo'
-                }
-              }
-            }
-          }
-        },
-        type: 'object',
-        properties: {
-          foos: {
-            type: 'array',
-            items: {
-              $ref: '#/definitions/foo'
-            }
-          }
-        }
-    } as JsonSchema;
-    const result = resolveSchema({schema: schema, dropPoints: {}} as ItemModel,
-      '#/properties/foos/items');
-    t.true(isItemModel(result));
-    t.deepEqual((<ItemModel>result).schema, {
-      definitions: {
-        foo2: {
-          type: 'object',
-          properties: {
-            bar: {
-              type: 'array',
-              items: {
-                $ref: '#'
-              }
-            }
-          }
-        }
-      },
-      type: 'object',
-      properties: {
-        bar: {
-          type: 'array',
-          items: {
-            $ref: '#/definitions/foo2'
-          }
-        }
-      }
-    });
-});
